Refactored the Controller and Payments model, mostly to handle low stock stores and Retails Express, where the the Order id may not find the original order created before we jump
There are also a few more setting in the Admin, logging can be disabled, and the store can set the return's success URL, which some onepage checkout modules require.

We use the magento redirect hook, which has the side effect of creating the order before the rediect is actioned, this is now deleted if there is an error or if force new order is set, we use the quote id for the merchant ref so we can create the order on return, and if the customer never returns there are no order hanging around.

As always feedback and suggestions are very welcome